requirements.txt
django
jsonschema
jsonref
djangorestframework
markdown
django-filter
